refactor to functions
